Commit to fix problems with paradynd on UltraSparc machines running Solaris2.6
	This is intended as a *partial* fix - it addresses *some*
(but *not* all) of the known problems with the sparc-solaris as
applied to solaris 2.6 on UltraSparc.  The idea is to allow
people to try to migrate to using the Solaris2.6 UltraSparc
machines - without trying to handle all the new UltraSparc
instructions and type migration issues.

	I changed the dyninstAPI and paradynd to the point where
they come up on solaris2.6, and work for simple things in our test
applications : namely - instrumenting user code in general, and the
library functions used in the I/O and synch metrics, and those
automatically instrumented by paradyn in DYNINSTinit.  There are still
some libc functions (e.g. _malloc_unlocked) which are not instrumented
correctly.  In general, the changes *should* bring paradynd up
to the point where if you exclude libc, you should be ok.

	Anyway, here is the list of known problems, both addressed
and not addressed:

ADDRESSED;
  - multiple instpoints found at the same address
    causes (newly-introduced) assertion failure
    example instance: ? (one of the standard libraries)
   fixed
 - restore with add side-effect
   fixed (see e.g. _exithandle)
 - calls designed to set O7 being marked as call sites (in non-relocated functions)
   fixed
 - unsupported return type warning (ret to call address + 16)
   fixed
 - I/O metrics broken
   fixed
 - instrumenting memory allocation breaks app
   bug # 1 - instrumenting _smalloc
    fixed (problem instrumenting _smalloc)
   bug # 2 - instrumenting _malloc_unlocked causes infinite recursion
    ???? - problem occurs in _malloc_unlocked (see below) - why????
    NOT fixed (don't try to instrument _malloc_unlocked)


NOT ADDRESSED:
 - "impdep" instructions:
 - nonoptimal "restore;retl" return
 - Type migration: 64-bit Address and Register types
 - 64-bit addressing instructions?
 - General paradyn problem that functions are identified by function_name, not
    by module_name/function_name can cause paradyn to instrument libc fork
    when libthread fork is the fork which is actually used - this can cause
    paradyn to miss calls to fork for some multi-process applications.