Changes in this commit implemented a dynamic code generation optimization. The basic idea is, all the definitions (including instrumentation point and instrumentation definition) of an instrumentation variable are grouped together, and form a primitive metricDefinitionNode.  The variable could either be a constraint variable or be a metric variable.  When a new variable is created, its definitions are checked against those of the variables already created, and reuse the old one if a match is found.
An example that shows the benefit of this optimization would be:
if the same focuses (a couple of procedures for example) are selected for
two different metrics that are of the same constraints, then they could
share the use of the constraint flag instead of creating two instances.
Of course the optimization is much general than this example.