Hello all:
This commit makes DyninstAPI functional on Win2k and
WinNT 4.0.

interesting bits:

: The compiler now takes the -MD option
which produces a multithreaded dll.  This is done
to ensure memory allocatd by the dyninstAPI.dll is
deallocated correctly by the mutator.

: There is a description in pdwinnt.C detailing the differences
between the NT4.0 debug api and the Win2k debug api,
and how the code has changed to handle them.

: The runtime dll libdyninstAPI_RT.dll now uses DLLMain()
to initialize the necessary data, which saves a bit
on the startup instrumentation, this is detailed in
pdwinnt.C as well.



enjoy.
chadd