Three changes:
1. Initialization of o7_live to false at a position that was missed before

2. I have changed the retun from a base tramp. Initially it was call insn
for every return for otherPoint but it is not right. If o7 is live
now it returns with save/call/restore as in leaf levevl functions.

3. I have seen that we did not check the offset and create a branch instruction
and sometimes the offset we put was too large so program went to paths it should not.
That happened when we relocate branch insn but had to create a 3 word space
as a branch target that calls back the original code.
I have cleaned that part and have added an assert check.