This commit enables arbitrary instrumentation points at the last instruction of a basic block by forcing function relocation (only when this type of inst point is requested) and inserting an appropriate NOP.  Additionally, since the inserted NOP is added inside the mini-tramp, the length of the basic block (as reflected by BPatch_basicBlock::getAddressRange) remains unchanged.
Also, to better control the automatic relocation that happens mostly "behind
the scenes" a new public method, BPatch::setAutoRelocation, has been added
(automatic relocation is on by default to preserve dyninst's current
semantics).  This method is not retroactive and it operates at the granularity
of a snippet (that is, creating inst points will succeed, but calls to
BPatch_thread::insertSnippet will fail if a function requires relocation but
it is turned off).