This commit removes the #ifdef linux to allow the previous fix to apply to all platforms.
The previous fix was for test4 subtest 3 on linux. If a function
was relocated in a function, then the process called exec() (without fork())
the function may still be marked as relocated in the execed process,
even though it has been reverted back to its original form.  Subsequent
attempts to relocate the function will fail (since Dyninst thinks it is
already relocated).