This should fix the problem with pardynd where starting a metric before starting the application accidentally continues the process (many thanks to Drew for figuring out exactly where this error was before I had even looked at it).
Also gets rid of (BPatch_callBefore BPatch_exit) error reports on linux,
but does not solve the larger problem.  Simply put, Dyninst does not
export anything that will create instrumentation at callPostInsn.  I
appreciate all of your input on this, but am still (slowly -- no hurry)
figuring out the best way to solve this.  Be it to account for every inst
point in paradynd, or to try to find a way that Dyninst can support
callPostInsn in a "natural" (and safe) way.

Adds BPatch_point->monitorCalls(), and the related functions.  Really not
much real code change here... just takes existing code for
MonitorCallSites and remaps it into the bpatch layer.

Adds a test case in test1 for this (test #40 -- monitor call sites).