This commit gets save the world working on AIX.  There will be another commit coming soon to refine this commit, refactor some code, make the code more robust, but this commit should work for all our test cases and as such will alert people when they have changed the code base and broken some part of save the world on AIX.
This commit does three things:

First, the DYNINST_tramp_guards array is allocated by
mmap on AIX (see note in RTcommon.c for a full explaination)
to allow the mutated binary to reload the array in the correct
place more reliably.

Next, the memory allocated in inferiorRealloc is now tracked
by the save the world data structures, so that this memory is
correctly saved and restored. (this code will be refactored to
make it easier to update the save the world data structures
whenever memory is allcoated in the mutatee).

Finally, when the save the world code is run the multiTramp at the
entry point of main has the call to DYNINSTthreadIndex() turned off
since this multiTramp is used to load the RT shared library, which
contains said function. This will be refined to only turn off the
call to DYNINSTthreadIndex() for the baseTramp in the multiTramp
that calls dlopen() to load the RT lib.