i(1) in decodeEventIfDueToIRPC we create a RPCSignal event if the PC is     at a relevant RPC address. (2) we save the _index_ of the rpc, ie its position in the allRunningRPCs_     vector. (3) Later, in handleRPCEvent, we use this index to look up the relevant     rpc, so that we don't have to search for it again.
In a single thread app, (2) and (3)  will occur in a predictable sequence,
without any change in the underlying data structures.

I think its possible, however, in mt apps, that, sometimes the
allRunninigRPCs_ vector will be modified _between_ (2) and (3), thus
invalidating the index that we saved.

The easy-but-ugly solution is to not rely on the index but to search for
the relevant rpc (again) in handleRPCEvent.