For the most part all the changes here are behind-the-scenes and disabled in the code (part of the build, but excluded from runtime).
Some notable changes:
(1)  pdvector is now almost fully deprecated -- it can be re-enabled by
activating the build flag -Dcap_use_pdvector.  Still active on AIX systems
since making the change to std::vector causes segfaults and our AIX
toolchain currently lacks a reasonably operational debugger...

(2)  I'm staging some new tests in what I'm calling "test0", in the
middle, not the newest testsuite.  I needed to bring some new tests online
and didn't want to step on any toes vis a vis the new testsuite, which as
I gather, is still in transition.  I consider test0 to be a temporary home
until we resolve issues regarding how we want to move forward with testing
sublibraries....  in this case elements of libcommon and libsymtabAPI, but
I suppose that all sublibs should be part of the consideration.

Like the majority of this commit, the contents of test0 are not really
relevant yet since most of the tested functionality is disabled -- though
I eventually will have some basic non-serialization-related tests of
symtabAPI running there in the nightlies.