This commit adds a new layer of abstraction to the SymtabAPI, and updates the DyninstAPI to use the new layer.
We now have two layers of representation for symbol information. They
are:

The Symbol class. This class represents a single symbol table entry,
consisting of the following:
  Names (mangled, pretty/demangled, typed)
  Address
  Symbol type (function, object, etc.)
  Linkage

The aggregate classes. Currently, these are the Function and Variable
classes. Each aggregate represents the collection of Symbols that
represent a single logical object, such as a function or
variable. Currently we define this as follows: all Symbols with the
same type and address are included in a unique aggregate.

There is still work to be done:

and Variable objects. Similarly, the current
addName/setSize/etc. methods should operate instead on aggregate
classes.

table.

---

Per-file commits:

SymtabAPI directory:
  Function.[Ch], Variable.[Ch]: headers and implementations for new
code.
  Symtab.*: functionality split into separate files for ease of
reading.
  annotations.[Ch]: Annotation definitions in their own file.

DyninstAPI directory:
  We now use Function and Variable objects where possible. There are
still some low-level Symbol lookups; these should be avoided if
possible but occasionally make sense (e.g., searching for "DYNAMIC").