Introducing recursive traversal parsing into Dyninst.
Instead of blindly parsing past call instructions, we now
parse the callee first, attempting to determine (if possible)
whether the function returns. If we can safely determine that
it _does not_, parsing is terminated at the call instruction
to prevent misparsed code.

This commit also disables speculative parsing of stripped
binaries using our current heuristics, which are known to
be faulty on some platforms.

A couple of other fixes crept in:

 - Functions created based on bad symbol table data
   (e.g., invalid addresses) are no longer parsed

 - A couple of build errors in the newtestsuite