Bugfix for test5_1 crash on Windows.
We were allocating a scratch virtual register in order to store the address of an indirect call.  We were then mapping this virtual register number to the physical register with the same ID.
This would enable us to overwrite ecx (call arguments), edx (ditto), ebp (bad for obvious reasons), and esp (really bad for obvious reasons) with the address that we were calling.
Fix: as per everywhere else in IA32 code generation, we use a known safe physical register (in this case, EAX).