Fixed a pernicious AMD64 crash bug with 32-bit mutatees.
When we load a library, we do a trap and a setPC deep in the guts of dlopen.  On AMD64, with a 32-bit mutatee, we were incorrectly using 64-bit values to set the mutatee PC.
Most of the time, the high 4 bytes were all zeros, and we were okay, but if they were ever non-zero, we'd segfault as soon as we tried to execute the next instruction.