This commit allows us to return a correct set of types in BPatch_type::getComponents() for types whose base type (constituent type) is a structure/class.
In Dyninst::Symtab::Type::getComponents(), we were ignoring all derived types that are not typedefs. However, pointers to classes/structures are derived types that should be handled correctly. Instead of just dealing with typedeftypes, we now treat all derived types the same: return the components of the base type.

We never created BPatch_type objects for some Dyninst::Symtab::Type's. Therefore, the upPtr in Type was set to null. However, for derived types, Dyninst::Symtab::Type::getComponents() relies on correctly set upPtr's. Now, whenever we create a new BPatch_type from a Type object that is a derived type, we create a BPatch_type object for the base type as well. upPtr's are set accordingly in BPatch_type constructor. For now, we are not registering these new BPatch_type objects anywhere else.