This should fix the crash in test 1_29 on Windows.
We had several bugs in our decoding of Ap-style operands on x86.  Since these only show up in far jumps and calls, this was not generally an issue.

I've added an InstructionAPI test that verifies that our decoding of far calls is correct, and it passes.  This should fix Windows.