Fix for test1_33 on Windows.
On Windows, we subtract off the load address from all of our symbols and code references before they leave Symtab.

When we parse a jump table, we get addresses to parse that do not have the load address subtracted off.  This makes that adjustment for both the table address and the table entries as needed.  Jump tables emitted by VC2003 are now parsed correctly.