Add a new "Interval Tree" type. This is a templated version of our codeRangeTree data structure used in Dyninst.
IntervalTree<class K, class V>

K is the datatype for lower and upper bounds. It must have a <, >, and
== operator defined. The ranges are stored as [lB..ub).

V is the value. It must have a default and copy constructor, as well as
an equality/inequality operator.

This class assumes the incoming ranges are disjoint; very, very, very
weird things will happen if this assumption is not held. It's also
not checked, so be careful in your usage.

One implemented optimization: joint ranges with identical types are
merged.

One exception: overlapping ranges with the same value are allowed
(needed to handle overlapping basic blocks).