Fix problem where multiple crashing mutators could send newtestsuite into infinite loop.
The newtestsuite resume log system had a problem where multiple if three mutators would crash in a row, it could forget about the first crash and attempt to rerun that test.  That led to an infinite loop as the newtestsuite would keep attempting to rerun the same crashing tests over and over.