Fix for bug 935 (test1_41 failure on Solaris)
When adding line information, we were using the
source file name instead of module name. In this case,
we were using test1_41_mutatee.c instead of
solo_mutatee_boilerplate.c

When we try to look it up, we were not able to find module
test1_41_mutatee.c as it does not exist and hence the test was failing.

This commit fixes the problem by using module name, and not
source filename, to add line information in Object-elf.C (parseStabLineInfo)