(1) Added -f <lib_name:func_name> parameter to parseThat. parseThat generates some simple instrumentation by default. With -f parameter, user can specify a function in a library to use as instrumentation. This will be useful to test binary rewriter.
(2) Fix for test_thread_8 race condition.
test_thread_8 uses global variable sync_test and async_test to
pass the thread_id. If the order of thread execution changes, the
test fails. Hence, this test is broken. The thread_id must be passed as parameter to the function
and cannot be a global variable.