IA64 fix for binary rewriter crashes
Most tests were crashing in IA64 because Object-elf was
not determining the dynamic section of a binary correctly.
We were using offset of the dynamic section in find its ELF
section number. But libc has two sections in the same offset as
the dynamic section. The first section is tbss of size 64 and
type NOBITS, followed by the dynamic section itself. We tried reading
the contents of tbss and intepreted it as dynamic section - and hence
crashed.

Now, we use a prioritized combination of offset, size and section type to
determine the dynamic section if there are more than one section
in the same offset.