enable a bunch of tests in rewriter mode.  add a couple new tests for the rewriter (snip_change_shlib_var and snip_ref_shlib_var) to test whether rewritten code that references and/or changes variables in a shared library will work properly.
changes the build structure for parseThat slightly to make it possible to add it to the
general overall dyninst build.  This was done to ensure that no race conditions could happen
with configure if, during the nightlies, several machines are running configure at the
same time.

Adds a parseThat test to the newtestsuite (with a hopefully hand ParseThat wrapper class
to make issuing calls to parseThat fairly easy)...  test_pt_ls.