This commit fixes bug 1008.
We were unable to obtain correct return type of functions when the return type attribute (DW_AT_type) only existed in the tag related to the declaration of the function and not in the tag related to the implementation. We now go and read that attribute from the tag that declares the function.