testsuite:  expand DYNINST_REWRITER_NO_UNLINK such that unique filenames are produced for each rungroup, this ensures that binaries are not overwritten by each successive group.  For a given rungroup, only a subset of the rewriter tests will be valid.  This also creates a report file, one per rewritten binary, that contains the list of tests that were implemented in that particular binary.
Also, if, in a rungroup, one test fails or crashes, all tests in the group were marked with failure.  This changes that:  now if a rewritten mutatee crashes or reports failure, the rungroup is broken apart into individual tests, with each being run seperately to provide a finer grain of reporting on what succeeded and what failed.