Rewriting static binary
This commit handles rewriting static binaries. We determine if
a binary is static in ObjectElf using PT_INTERP. We use this
condition to selectively not execute code related to dynamic linking
and relocation in ObjectElf and emitElf.

We also disable traps in binaryEdit and baseTramps - since static
binaries cannot load our runtime library.


We still do not handle stripped binary in this commit - but thats
coming soon.