Fix for parseThat build failures
parseThat build involves executing parseThat/Makefile and
parseThat/$PLATFORM/Makefile. We were setting
TO_CORE (path to dyninst directory) as "../.." which is correct
if we are in parseThat/$PLATFORM/Makefile but
wrong if we are in parseThat/Makefile.

The only reason to use TO_CORE in parseThat/Makefile is to set the
platform if not set, using scripts/sysname. This commit fixes this
problem by setting the path to scripts correctly, without using TO_CORE.