Fixes for bug 1028:Single snippet removal at multi-inst point
In this fix, we regenerate multitramp when its basetramp is changed.
A new function called updateCode is added to multiTramp class. This function
regenerates new code for the multiTramp when a miniTrampInstance is
removed.
I also fixed a bug in multiTramp.C - removeCode function - condition to
check if a particular multiTramp is the current version mapped in the process
table.