AIX fixes
We were using /proc/pid/object/a.out names for creating Symtab Objects.
When we do an exec, the new execed file is also called
/proc/pid/object/a.out and hence Symtab does not re-parse it.

This is fixed by passing the executable name itself instead of a.out

The number of "createMappedFiles" and "closeMappedFiles" is not matched
properly. This was causing some segfaults. This is also fixed.