Fixed a regression when loading a library via the openResolvedLibraryName function.
I introduced a bug where the search to locate a library by name ended after
finding the first possible candidate for the library. However, on platforms
that have different ABI versions of the same library such as x86_64 Linux, this
behavior is incorrect because the 64-bit version could have been found when the
user was looking for the 32-bit version. The fix was to gather all possible
candidates for the library, iterate through these candidates in an order
dictated by standard convention, and select the first library that can be
opened successfully and matches the ABI of the original binary.