Fixed problems related to loading libraries introduced by a previous commit.
1) On platforms that do not support the binary rewriter, opening a binary with
BPatch_binaryEdit would fail because openResolvedLibraryName was not
implemented for these platforms. A generic version of openResolvedLibraryName
was implemented on these platforms. openResolvedLibraryName is now used to open
the RT library because it can be a collection of BinaryEdits in the static
binary case. Previously, the RT library was opened with openFile.

2) Prior to this commit, in order to work with a static binary via
BPatch_binaryEdit the DYNINSTAPI_RT_LIB environment variable needed to point
to the .a version of the RT lib. This inhibits a user from working with a
dynamic binary and a static binary at the same time. Now, the
DYNINSTAPI_RT_LIB variable is used to find the directory that holds both
versions of the RT library. The correct version is chosen depending on the
type of binary opened via BPatch_binaryEdit.