Extend testsuite to build and test static binaries.  Adds new dimension to test space -- dynamic vs. static link. runTests and test_driver now accept the options(-dynamiclink,-staticlink) to run against dynamically or statically linked binaries. The default is to have both options on, but statically linked binaries will only be tested in rewriter mode. This included some changes to the test infrastructure code as well as the specification.
Also, fixes some bugs in the rewriter of static binaries: restoring symbol
information after performing the rewrite to all Symtab caching to work correctly
and automatically loading the dependencies for the static RT library (just
libc for now).

All rewriter tests pass for static binaries except test5_6. test5_6 (Exception
test) instruments the catch block of an exception block. The mutatee makes a
call to a function that always throws an exception; parsing currently
classifies this case as a non-returning function. In the calling function, the
non-returning call causes the parser to stop parsing the function and never
find the catch block. Because of this, the mutator fails to find and instrument
the call block. This isn't a problem for the dynamic case because the call is
routed through the PLT and is therefore, not a non-returning call. At the time
of the commit, I didn't know how to fix this.