Fixes parent calculation for natural loops
This commit fixes a bug in our creation of a hierarchy between
the natural loops in a function.  In particular, if there are
multiple natural loops with different back edges, but the same
set of basic blocks, we assign the loops to be each other's
parents in the hierarchy, creating a loopy hierarchy of loops.
With this commit, we no longer assign a loop to be a parent of
another loop of the same size.