Further refactoring of AddressTranslate for FreeBSD
A newly created process doesn't have the DT_DEBUG field populated, so
changes were made to lazily parse this field. However, the trap address
needs to be known immediately after an AddressTranslator is created.
Luckily on FreeBSD, there is a symbol in the dynamic linker for the trap
address.

If DT_DEBUG is unavailable, the trap address is parsed from the dynamic
linker. If DT_DEBUG is available, the trap address is parsed from the
r_debug structure.

Consequently, to parse the dynamic linker, we need to know its base
load address in memory. This can be retrieved from memory maps. This
has been implemented for FreeBSD and doesn't require procfs. This
required some refactoring of the existing memory map parsing support
on Linux.

On another note, plat_install in ProcControlAPI was factored to
sysv.C because the functionality is the same on both Linux and FreeBSD.