More changes to support multithreaded debuggees on FreeBSD.
Updates to FreeBSD Makefiles and the RT library with changes only made to 7.2
i386 so far.

Includes a function to determine the initial LWP of a running process.

Includes some reworking of the pc_thread_cont test. These changes will be
overridden by a future commit.

Made miscellaneous bug fixes throughout, work in progress to get iRPC tests
working for multithreaded debuggees on FreeBSD.

The implementation of thread stops and continues was reworked in this commit.
This commit removes the fields pending user continue and pending continue, and
it removes the EventContinue. FreeBSD requires that a process be stopped before
continuing a thread. So to continue a thread in a process with other running
threads, the process must be stopped before issuing the continue. To implement
this, the desync/restore operations on the internal state were used to stop a
process with running threads before continuing a running thread. This
implementation appears in both the threadPool continue and thread continue
functions.

Also encountered a weird bug when attaching to MT debuggees. On attach, all
threads are suspended. When continuing a single of these threads, multiple
threads run which makes ProcControl's state inconsistent with the OS's state.
This commit includes a workaround (see the FreeBSD specific handlers for more
info).