Added hybrid analysis modes, not fully functional.
This commit adds three analysis modes, BPatch_normalMode,
which is the default and the only mode to trigger at present,
BPatch_exploratoryMode, which causes runtime instrumentation
of unresolved indirect control transfers and thereby may
discover additional code at runtime (not fully functional
yet), and BPatch_defensiveMode, which assumes that the binary
is defensive, possibly self-unpacking and self-modifying.

Also, this commit should fix the Windows failures that popped
up this weekend after an upgrade to VS2008 project and solution
files.