Fixes Windows build errors cause by my last commit
In my last commit I failed to merge these fixes;
without them the Windows platform was not building.
The fixes are to code that only triggers on
the Windows platform, with the exception of the
mapped_object::isSystemLib function which I modified
to make it static so that we can determine whether an
object is a system library prior to parsing it.