Removes cyclic dependency between symEval and parseAPI
It turned out that the cyclic dependency I had introduced between
ParseAPI and symEval caused problems on other platforms than Windows.
These build problems only manifested when building from a clean
repository.  As of this commit, the parseAPI no-longer depends on
symEval, and we'll work on a way to allow for slicing to be allowed
in the ParseAPI without introducing cyclic library dependencies.