Bluegene rewriter changes
In object-elf.C,
We identify the platform as Bluegene using the comment section.
The comment section should contain the string BGP.

In Symtab.C,
We adjust the pagesize of the new PT_LOAD segment (if any) to 1MB
due to compute node restrictions.

In emitElf.C,
We use one of the following mechanisms to
add new loadable section in Bluegene:
1) We either expand BSS section
2) We replace NOTE section with a LOAD section.