Bad logic was preventing parsing of some jump tables.
This only applied where the indirect branch is at the end of the
first block of the function and no calls to that function had been
observed (so the block's sources() vector was empty). This bug
cropped up because functions with unparsed jump tables are now marked
uninstrumentable, and 64-bit libc's printf happens to be just such
a function, causing instrumentation that inserts calls to printf
to break.