Create new stack frame when getScratchRegister fails.
In power platforms, while relocating Thunk code or
emitting load and store shared code outside of baseTramp,
getScratchRegister may fail when there are no free registers.
When there are no free registers, we create a new stack frame,
push register values on the stack and create free registers.

After the operation, we tear down the stack and restore the original
values of the registers.