Fix for bug 1064 - instrumentation of malloc in libc causes seg. fault.
The relocations for inter-module variable references/function calls included in
instrumentation were being generated incorrectly for shared libraries that have
a non-zero load offset. A notable example of a library that has this
characteristic is libc. The problem was the address stored in the relocation
was decreased by the library's load offset. This is incorrect because the
address stored in a relocation for a shared library is a base virtual address,
not an offset.