Initial support for FreeBSD.
This commit serves as a marker for when the initial FreeBSD changes
were merged with git-head. Please see previous commits for
further descriptions of changes.

The initial support for FreeBSD includes a working binary rewriter
and an almost complete port of ProcControlAPI ( I have observed
some race conditions for the iRPC test when working on multithreaded
debuggees ).

It also removes the FreeBSD 8.0 build directories and support in the testsuite
for FreeBSD 8.0, as the initial support was only tested on FreeBSD 7.2.