Move isStaticExecutable method from BPatch_binaryEdit to BPatch_addressSpace.
The isStaticExecutable method isn't specific to offline editing so a
logical place for this method is in the parent class. Additionally,
this change will not break existing code (although there shouldn't be
any references to this function in user's code because this method doesn't
exist in any released version of Dyninst).

This commit also changes any references in the testsuite to be more generic.