Re-wrote code that keeps memory in sync with mapped data
We now update if there's a control transfer target on an
unprotected page, the underlying bytes have changed, and
we haven't updated since the last run.  Also made some
fixes in symtabAPI so we can parse code that is not
in a .text region if the binary is in defensive mode.

and yet
thorough, an We now keep a list of pages whose code is protected
so we don't check them for updates