fixes IBSTree interval removal for 0-sized intervals
The IBSTree can contain intervals of zero length, and we weren't handling this corner case properly for interval removal.  Fixed.  We add zero-length intervals to the tree when tracking uninitialized sections of the program binary (i.e., sections that have zero-size in the binary but for which physical memory is allocated when the binary is loaded).

warning: this touches dynutil/h, so it causes just about everything to rebuild from scratch.