bug 1090: test1_20 BGP failure
Floating point control and status register
is 8 bytes long on ppc32_linux. They must be
saved at 8 byte boundary on the stack (as we
use stfd instruction). Our destination on the stack
was not aligned correctly. This caused an alignment
error and hence an exception on BGP. On other machines,
this would be a performance hit.

This commit fixes the alignment.