Code cleanup: Gap parsing flags were ignored
We use a flag to indicate whether gap parsing should be
on or off for a particular binary, but it was being ignored,
so that gap parsing was always on for platforms that implement
it.  This commit cleans up the code but shouldn't have any
practical effect, since gap parsing is still on by default for
all platforms that support it.