Fix for using the FreeBSD provided kernel patch for the lost signal kernel bug.
The patch provided by the FreeBSD developer fixed the problem for some
signals, but not all signal. Specifically, job control signals (SIGSTOP
and friends) and SIGKILL. This fix changes ProcControl to issue SIGUSR2
signals to stop threads instead of SIGSTOPs. There is no semantic
difference between using a SIGUSR2 to stop threads as compared to using
a SIGSTOP.

Also, fixes a bug in ProcControl testsuite mutatees introduced by a
previous commit. errno needs to be checked immediately after the
syscall that produced the error we care about.