Adds support for ppc64_linux linker stubs (SecurePLT)
The PPC64 architecture is limited to a 24-bit branch instruction.  To
circumvent this limitation, binutils linkers generate "stub" code close
to the original branch which, using multiple instructions, jump to the
actual target.

This confuses our current parsing scheme, causing it to generate a number of
unwanted "targXXXXX" functions.  While technically correct, this prevents the
ability to do callsite instrumentation.

This commit allows our parser to correctly recognize these targets as stubs,
determine their true target, and return it to the parser as if it were a
normal call.