Fixes for rewriting stripped, static binaries (bug 1113).
When rewriting stripped static binaries, heuristics are used to find the
functions that execute constructors and destructors for global objects.  These
functions are replaced with functions in the Dyninst RT library so additional
constructors and destructors in newly linked static libraries can also be
executed at initialization and finalization. The code to do this was broken.

Also, fixed error reporting for link-time errors when rewriting
static binaries.