Fixed cases where post-call padding not set
Was failing to set post-call padding jumps if our stack-analysis was inconclusive,
because we weren't instrumenting return instructions.  We also weren't padding when
parseAfterCallAndInstrument didn't result in new parsing (if the call was assumed
to be non-returning, but the subsequent instruction had been parsed anyway due to
a jump to that address).