Fixed signal handling bug and mapped_object size bug
finalizeInsertionSet was stopping already stopped processes; this had
a bad effect on the program if it was executing a callback from a signal
handler (in Windows) at the time of the instrumentation.  Also, a check
to see if we're writing within the bounds of a mapped_object failed for
programs that write to a process's uninitialized memory after
the end of the code.