fix handling of calls to invalid addresses
Call instructions to invalid addresses were previously thrown out as not being real call instructions, meaning that they did not terminate blocks and were considered not to return.  We now treat such call instructions in the same way as indirect calls; we end the block, parse at the fallthrough address, adding an unresolved call edge to the sink block.