Replaces existing process control in Dyninst with ProcControlAPI stubs.
Refactors BPatch, BPatch_process, and BPatch_thread to use the ProcControlAPI
stubs.  These stubs are implemented in the classes: PCProcess, PCThread, and
PCEventHandler.

This commit touches so many files because the process control was hooked deeply
into Dyninst. This commit removes a fair amount of references to process
control files that were unnecessary.

It is important to note that this commit represents a work in progress.