Power jump table parsing
1) Extensions to parseAPI(IA_powerDetails) to
find jump tables in ppc32 code.
The addi-addis combination instructions can
be anywhere in the function before the indirect
jump - not necessarily in the same block as the
indirect jump instruction.
The jump table start address is stored in got
section. For shared libraries, this is filled by
the dynamic loader via a relocation entry. Hence,
we should find the relocation entry corresponding
to the table start address to find the table in
rodata section.
2) symtabAPI: Relocation entries are added to
the corresponding DT_REL sections (rel.dyn and
rel.plt) in addition to the target region of the
relocations (plt section). This enables us to
get a list of relocations in a particular ELF
section.