Block split recording and snippet removal of executing functions.
We weren't reporting block splits resulting from a new function entry point splitting an existing block, fixed.
We can't add or remove code to a function if an overwrite loop is executing in the meantime, in which case we should fail.  Currently deleteSnippet returns false, but perhaps it's finalizeInsertionSet that should fail.
Also, there's some debug code here