Fix two bugs in Dyninst's exec handling code.
Both of these bugs were reported by the Open SpeedShop folks when
using Dyninst from the openssd daemon. They existed in both Dyninst 6.1
and GIT HEAD.

The first bug occurs when a multithread capable process exec's an executable
that is not multithread capable. We were incorrectly assuming that the exec'd
process was still multithread capable.

The second bug was observed when a multithread capable process exec's an
executable that is also multithread capable. The bug was that the tramp guard
AST was being cached across exec's and results in asserts in code generation
because the tramp guard variable created during mulitthreaded initialization
doesn't match the variable maintained by the tramp guard AST. The fix is to
clear the cached the tramp guard AST on exec.