Better return status labeling for shared funcs and Rose input sanitizing
If two functions share code, and function A parses into an already parsed function B,
we can be sure that B's return status is the same as function A's if B is contained
in A.  As of this commit, we assign B's return status to A if A parses into B's entry point.
This commit also sanitizes the pushf instruction so that it can be passed into Rose.