Use provided names for PLT functions, if available.
During parsing, calls through the PLT lead to creation of functions
for the PLT entry stubs. These functions were getting the standard
targ* names we give to recursively discovered functions, which names
can be highly annoying when you KNOW that the real name of the PLT
entry in question exists RIGHT THERE in the symbol table. This commit
does the Right Thing.