Force stack alignment on x86/x86_64 platforms.
Both x86 and x86_64 platforms require an aligned stack pointer at certain
points of execution.  This commit introduces a preamble in our state saving
code to forcably align the stack pointer.

Care is also taken to ensure the stack is aligned before we generate a
function call on x86_64.  Similar functionality may be needed for x86
in the future.