Optimize order of slice expansion
The order that slice nodes are visited in slice -> AST expansion matters
quite a bit; bad visitation order can lead to a great deal of extra work
being done as nodes are repeatedly visited to fill in partially expanded
ASTS. Because cycles in the slice DAG are removed prior to expansion,
an optimal ordering is possible. This commit implements that ordering. It
also correctly avoids traversing the `skip edges' that are removed to
break cycles in all instances of edge enumeration.