Fixes to call emulation, slicing.
Slicing inserts a "widen" at int3 instructions since these will invoke
the OS, and the PC will be visible on the stack, making it so that
PC emulation can't know that the return address is not used.  Also
fixed slicing to deal with call instructions that have no return
instruction.  Finally, emulated calls that get post-call padding
need jumps even if the called function is immediately after the call.