Eliminate spurious `thunk function' detection on x86
A `thunk function' is one that contains exactly two instructions:

mov (%esp),<a register>
ret

This construct is used in position independent code to retrieve
the program counter---see for example the __i686.get_pc_thunk.*
functions in glibc. Previously, any function that had a mov
instruction that referenced the stack pointer followed by a return
would be treated as a thunk. This commit properly checks for a
zero-offset dereference of the stack pointer.