Initial, somewhat working version of Dyninst with ProcControlAPI.
This commit represents a work-in-progress.

A fair amount of refactoring through out the BPatch classes that
use process control.

The majority of the implementation work went into the PC* classes,
they interact with ProcControlAPI. This commit introduces the
new threading model where a thread is used to execute ProcControl
callbacks and queue events for processing by the user thread. This
functionality is encapsulated by the PCEventHandler class. Most of
the events generated by ProcControlAPI are handled correctly. The
remaining events are Fork, Exec, ThreadCreate, ThreadDelete and
Signal.

Removed references in baseTramp to RPCs--ProcControlAPI automatically
saves and restores all registers when RPCs are executed.